+++ osu.Game/Screens/Menu/ButtonSystem.cs
@@ -126,7 +126,7 @@ namespace osu.Game.Screens.Menu
         {
             buttonsPlay.Add(new Button(@"单人游戏", @"button-solo-select", FontAwesome.Solid.User, new Color4(102, 68, 204, 255), () => OnSolo?.Invoke(), WEDGE_WIDTH, Key.P));
             buttonsPlay.Add(new Button(@"多人游戏", @"button-generic-select", FontAwesome.Solid.Users, new Color4(94, 63, 186, 255), onMulti, 0, Key.M));
-            buttonsPlay.Add(new Button(@"表格", @"button-generic-select", OsuIcon.Charts, new Color4(80, 53, 160, 255), () => OnChart?.Invoke()));
+            buttonsPlay.Add(new Button(@"排名", @"button-generic-select", FontAwesome.Regular.ChartBar, new Color4(80, 53, 160, 255), () => OnChart?.Invoke()));
             buttonsPlay.Add(new Button(@"一个神秘的按钮", @"button-generic-select", FontAwesome.Solid.Boxes, new Color4(80, 53, 160, 255), () => OnBeatmapListing?.Invoke() ));
             buttonsPlay.ForEach(b => b.VisibleState = ButtonSystemState.Play);